EY-2013 Egen fnr valideringsfunksjon for behandling
