+++ libs/etterlatte-token-model/src/main/kotlin/no/nav/etterlatte/token/Bruker.kt
@@ -6,9 +6,6 @@ sealed class Bruker {
     abstract fun ident(): String
 
     abstract fun accessToken(): String
-    abstract fun kanAttestereFor(ansvarligSaksbehandler: String): Boolean
-    abstract fun harRolle(rolle: String?): Boolean
-
     companion object {
         private fun erSystembruker(oid: String?, sub: String?) = (oid == sub) && (oid != null)
         fun of(
@@ -35,16 +32,12 @@ data class SystemBruker(val oid: String, val sub: String) : Bruker() {
     override fun ident() = Fagsaksystem.EY.navn
 
     override fun accessToken() = throw NotImplementedError("Kun relevant for saksbehandler")
-    override fun kanAttestereFor(ansvarligSaksbehandler: String) = true
-    override fun harRolle(rolle: String?) = false // Per no ikkje relevant for systembrukar
 }
 
 data class Saksbehandler(val accessToken: String, val ident: String, val jwtTokenClaims: JwtTokenClaims?) : Bruker() {
     override fun ident() = ident
 
     override fun accessToken() = accessToken
-    override fun kanAttestereFor(ansvarligSaksbehandler: String) = ansvarligSaksbehandler != this.ident()
-    override fun harRolle(rolle: String?) = jwtTokenClaims?.harRolle(rolle) ?: false
 }
 
 enum class Claims {
@@ -52,5 +45,3 @@ enum class Claims {
     oid, // ktlint-disable enum-entry-name-case
     sub // ktlint-disable enum-entry-name-case
 }
-
-private fun JwtTokenClaims.harRolle(rolle: String?) = containsClaim("groups", rolle)
